+++ instrumentation/persistent/src/OpenTelemetry/Instrumentation/Persistent.hs
@@ -1,14 +1,17 @@
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module OpenTelemetry.Instrumentation.Persistent (
   wrapSqlBackend,
+  wrapSqlBackend',
 ) where
 
 import Control.Monad.IO.Class
 import Control.Monad.Reader
 import Data.Acquire.Internal
+import qualified Data.HashMap.Strict as H
 import Data.Maybe (fromMaybe)
 import Data.Text (Text)
 import qualified Data.Text as T
@@ -51,7 +54,7 @@ lookupOriginalConnection :: SqlBackend -> Maybe SqlBackend
 lookupOriginalConnection = Vault.lookup originalConnectionKey . getConnVault
 
 
-connectionLevelAttributesKey :: Vault.Key [(Text, Attribute)]
+connectionLevelAttributesKey :: Vault.Key (H.HashMap Text Attribute)
 connectionLevelAttributesKey = unsafePerformIO Vault.newKey
 {-# NOINLINE connectionLevelAttributesKey #-}
 
@@ -62,15 +65,28 @@ connectionLevelAttributesKey = unsafePerformIO Vault.newKey
 wrapSqlBackend ::
   MonadIO m =>
   -- | Attributes that are specific to providers like MySQL, PostgreSQL, etc.
-  [(Text, Attribute)] ->
+  H.HashMap Text Attribute ->
   SqlBackend ->
   m SqlBackend
 wrapSqlBackend attrs conn_ = do
   tp <- getGlobalTracerProvider
+  pure $ wrapSqlBackend' tp attrs conn_
+
+
+{- | Wrap a 'SqlBackend' with appropriate tracing context and attributes
+so that queries are tracked appropriately in the tracing hierarchy.
+-}
+wrapSqlBackend' ::
+  TracerProvider ->
+  -- | Attributes that are specific to providers like MySQL, PostgreSQL, etc.
+  H.HashMap Text Attribute ->
+  SqlBackend ->
+  SqlBackend
+wrapSqlBackend' tp attrs conn_ =
   let conn = Data.Maybe.fromMaybe conn_ (lookupOriginalConnection conn_)
-  -- TODO add schema to tracerOptions?
-  let t = makeTracer tp "hs-opentelemetry-persistent" tracerOptions
-  let hooks =
+      -- TODO add schema to tracerOptions?
+      t = makeTracer tp "hs-opentelemetry-persistent" tracerOptions
+      hooks =
         emptySqlBackendHooks
           { hookGetStatement = \conn sql stmt -> do
               pure $
@@ -83,7 +99,7 @@ wrapSqlBackend attrs conn_ = do
                                 t
                                 ctxt
                                 sql
-                                (defaultSpanArguments {kind = Client, attributes = ("db.statement", toAttribute sql) : attrs})
+                                (defaultSpanArguments {kind = Client, attributes = H.insert "db.statement" (toAttribute sql) attrs})
                             adjustContext (insertSpan s)
                             pure (lookupSpan ctxt, s)
                           spanCleanup (parent, s) = do
@@ -104,7 +120,7 @@ wrapSqlBackend attrs conn_ = do
                             )
                             (stmtQueryAcquireF f)
                   , stmtExecute = \ps -> do
-                      inSpan' t sql (defaultSpanArguments {kind = Client, attributes = ("db.statement", toAttribute sql) : attrs}) $ \s -> do
+                      inSpan' t sql (defaultSpanArguments {kind = Client, attributes = H.insert "db.statement" (toAttribute sql) attrs}) $ \s -> do
                         annotateBasics s conn
                         stmtExecute stmt ps
                   , stmtReset = stmtReset stmt
@@ -112,7 +128,7 @@ wrapSqlBackend attrs conn_ = do
                   }
           }
 
-  let conn' =
+      conn' =
         conn
           { connHooks = hooks
           , connBegin = \f mIso -> do
@@ -123,16 +139,16 @@ wrapSqlBackend attrs conn_ = do
                       Just ReadCommitted -> " isolation level read committed"
                       Just RepeatableRead -> " isolation level repeatable read"
                       Just Serializable -> " isolation level serializable"
-              let attrs' = ("db.statement", toAttribute statement) : attrs
+              let attrs' = H.insert "db.statement" (toAttribute statement) attrs
               inSpan' t statement (defaultSpanArguments {kind = Client, attributes = attrs'}) $ \s -> do
                 annotateBasics s conn
                 connBegin conn f mIso
           , connCommit = \f -> do
-              inSpan' t "commit" (defaultSpanArguments {kind = Client, attributes = ("db.statement", toAttribute ("commit" :: Text)) : attrs}) $ \s -> do
+              inSpan' t "commit" (defaultSpanArguments {kind = Client, attributes = H.insert "db.statement" (toAttribute ("commit" :: Text)) attrs}) $ \s -> do
                 annotateBasics s conn
                 connCommit conn f
           , connRollback = \f -> do
-              inSpan' t "rollback" (defaultSpanArguments {kind = Client, attributes = ("db.statement", toAttribute ("rollback" :: Text)) : attrs}) $ \s -> do
+              inSpan' t "rollback" (defaultSpanArguments {kind = Client, attributes = H.insert "db.statement" (toAttribute ("rollback" :: Text)) attrs}) $ \s -> do
                 annotateBasics s conn
                 connRollback conn f
           , connClose = do
@@ -140,7 +156,7 @@ wrapSqlBackend attrs conn_ = do
                 annotateBasics s conn
                 connClose conn
           }
-  pure $ insertOriginalConnection conn' conn
+   in insertOriginalConnection conn' conn
 
 
 annotateBasics :: MonadIO m => Span -> SqlBackend -> m ()