+++ instrumentation/yesod/src/OpenTelemetry/Instrumentation/Yesod.hs
@@ -29,6 +29,7 @@ module OpenTelemetry.Instrumentation.Yesod (
   handlerEnvL,
 ) where
 
+import qualified Data.HashMap.Strict as H
 import Control.Monad.IO.Class (MonadIO)
 import Data.List (intercalate)
 import Data.Map (Map)
@@ -328,14 +329,15 @@ openTelemetryYesodMiddleware rr (HandlerFor doResponse) = do
   mspan <- Context.lookupSpan <$> getContext
   mr <- getCurrentRoute
   let sharedAttributes =
-        catMaybes
-          [ do
-              r <- mr
-              pure ("http.route", toAttribute $ pathRender rr r)
-          , do
-              ff <- lookup "X-Forwarded-For" $ requestHeaders req
-              pure ("http.client_ip", toAttribute $ T.decodeUtf8 ff)
-          ]
+        H.fromList $
+          catMaybes
+            [ do
+                r <- mr
+                pure ("http.route", toAttribute $ pathRender rr r)
+            , do
+                ff <- lookup "X-Forwarded-For" $ requestHeaders req
+                pure ("http.client_ip", toAttribute $ T.decodeUtf8 ff)
+            ]
       args =
         defaultSpanArguments
           { kind = maybe Server (const Internal) mspan