+++ instrumentation/cloudflare/ChangeLog.md
@@ -1,3 +1,7 @@
 # Changelog for cloudflare
 
 ## Unreleased changes
+
+### Breaking changes
+
+- Use `HashMap Text Attribute` instead of `[(Text, Attribute)]` as attributes