+++ instrumentation/persistent/ChangeLog.md
@@ -1,7 +1,11 @@
 # Changelog for hs-opentelemetry-persistent
 
+## Unreleased changes
+
+### Breaking changes
+
+- Use `HashMap Text Attribute` instead of `[(Text, Attribute)]` as attributes
+
 ## 0.0.1.0
 
 - Initial release
-
-## Unreleased changes